ocb3: use `dbl` crate
Replaces the vendored implementation of GF(2^128) doubling (i.e.
multiply-by-x) with the `dbl` crate.

Also moves the remaining functions in the `util` module into lib.rs.
