+++ ocb3/Cargo.toml
@@ -19,6 +19,7 @@ rust-version = "1.72"
 aead = { version = "=0.6.0-pre.0", default-features = false }
 cipher = "=0.5.0-pre.4"
 ctr = "=0.10.0-pre"
+dbl = "=0.4.0-pre.4"
 subtle = { version = "2", default-features = false }
 zeroize = { version = "1", optional = true, default-features = false }
